refactor:调整权限部分的目录结构，剥离mybatis-plus的实现目录，使得业务结构更清晰

+++ boot/kewen-auth-web-starter/src/main/java/com/kewen/framework/boot/authority/advance/authedit/DefaultAuthDataEditBusiness.java
@@ -11,12 +11,12 @@ import lombok.Data;
 @Data
 public class DefaultAuthDataEditBusiness implements AuthDataEditBusiness {
 
-    private Integer id;
+    private Long id;
     private AuthorityObject authority;
 
 
     @Override
-    public Integer getBusinessId() {
+    public Long getBusinessId() {
         return id;
     }
