+++ boot/kewen-auth-web-starter/src/main/java/com/kewen/framework/boot/authority/model/BusinessAuthorityEditReq.java
@@ -15,7 +15,7 @@ import javax.validation.constraints.NotNull;
 public class BusinessAuthorityEditReq {
 
     @NotNull
-    private Integer businessId;
+    private Long businessId;
 
     @NotBlank
     private String module ;
